pyproject.toml
[project]
name = "woogent"
version = "1.0.0"
description = "WooCommerce Universal Commerce Protocol (UCP) API + Gemini Demo"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.6",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy>=2.0.36",
    "aiomysql>=0.2.0",
    "pydantic>=2.10.4",
    "pydantic-settings>=2.7.1",
    "python-jose[cryptography]>=3.3.0",
    "cryptography>=43.0.3",
    "python-multipart>=0.0.20",
    "httpx>=0.28.1",
    "gradio>=5.9.1",
    "google-generativeai>=0.8.3",
    "python-dotenv>=1.0.1",
    "stripe>=14.4.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.24.0",
]

[tool.uv]
package = false
